Fix broken test
The previous commit fixed that the `assert_no_selector` method was not
actually doing any assertions. This revealed that this test was actually
failing as the rendering was producing the HTML that we were trying to
assert was absent.

To fix this I've had to modify the content item to have the links that
indicate it is part of a step-by-step nav.