Only include capybara helpers when needed
The previous commits contain fixes for when Capybara assertions were
mixed into tests that didnt actually make use of the Capybara DSL and
thus were not actually asserting correctly.

To try prevent this happening again I've configured the Capybara helpers
(and thus assertions) to only be available to tests that are tagged with
`:capybara`. This will mean that trying to use `assert_no_selector` will
raise a NoMethodError rather than having a test that passes with
whatever is given.

Ideally we'd not have had to put a tag of `capybara` on and instead
group this just to specs for features (those in /specs/features) however
in this test suite there isn't a clear test pyramid distinction and the
browser feature specs are frequently mixed in with more regular unit
tests. This meant taking the more clunky tag approach was needed.

I've also fixed tests that were broken due to reliance on Capybara
assert methods.