Add Capybara::DSL to visual_regression tests
It doesn't strike me that they'll ever be a scenario where we want
visual regression tests without capybara. Therefore I've added the
visual_regression tag as one to also use the Capybara::DSL.

While doing this I took the liberty to update the visual_regression tag
usage to be the short form.