Remove `GOVUK_DISABLE_SEARCH_AUTOCOMPLETE` feature flag
We are centralising the enabling/disabling of autocomplete in Search API
v2 so we don't need individual feature flags across several apps and in
the component itself.

The rationale for being able to disable autocomplete centres on a
scenario in which the underlying search engine returns unsuitable
results. The `search_with_autocomplete` component degrades gracefully
anyway when no suggestions are returned, so we are moving to a simpler
setup where there is a feature flag on Search API v2 to force returning
empty responses instead.