GA4 search tracker: Normalise initial value
In the search tracker, we do not want to fire the search event if the
keyword(s) are unchanged.

The way we do this is by comparing the current value of the search field
to its initial value, but there was a subtle bug: we normalised the
current value on every check, but not the initial value when it was
being set.