Make super-layout-nav toggle button border bottom transparent
This fixes the toggle button in its unhovered/unfocused state having a 1px dark navy bottom border. The colour change was made in https://github.com/alphagov/govuk_publishing_components/pull/2413 and the colour didn't get updated when the header design subsequently changed.

Change the bottom border to have a transparent colour, instead of removing it so that when user changes the default colours of the browser such as in Firefox, or uses Windows High Contrast mode, the button still has a bottom border.

The change doesn't impact the current homepage header which continues to set its own brighter blue bottom border.