Update styles dependency in super navigation component
The super navigation layout header component is dependant on the styles from the `header` component in the design system.

This change removes the call to `add_gem_component_stylesheet("layout-header")` and imports the styles required from the design system directly.

This should help make it clear which styles are required to render the super navigation layout header component and reduce the overall CSS size, as the layout header component from the gem includes lots of other unused styles, search, skip-link, tag for example.