Fix search button width
The search icon width is set to 21px, and has 15px of left and right padding by default, which adds up to 51px.

The large-navbar variation of the search toggle button has 24px of padding to the left and right, which adds up to 69px.

Updated the SASS variable name to more accurately reflect what it is used for.