Add an aria-label to the devolved nations component
The devolved nations component currently uses the section tag but doesn't have an aria-label. This might have been an oversight since it means the component doesn't get announced as a landmark region by screen readers. Add an aria-label to correct this, the value "Nation" was agreed with our content designer.

The wider context is that the notice and summary banner components are currently set as landmark regions as they are deemed to contain important information that it might be useful for the user to know about when they inspect the outline of the page. To be consistent, it makes sense to make devolved nations into a landmark region as well. (There has been some conversation though about whether some of these components could be consolidated to reduce the number of landmark regions and to generally help to to improve the information hierarchy on document pages. Our team have tickets in our backlog to look at wider related improvements for this.)