`Ga4FinderTracker` extendability functionality
- Change `trackChangeEvent` to use `event` as a parameter instead of
`eventTarget`. This means that `trackChangeEvent` can now be added
directly to an event listener.

- Take the `changeEventMetadata` from the closest element that has the
data attribute `data-ga4-change-category`. This makes it possible to
add `trackChangeEvent` directly to an event listener.

- Replace lines that duplicated the same functionality as the function
already present within `GOVUK.analyticsGa4.core`.

- `index` was being assigned in a way that meant that the resulting event
object would be assigned `[Object object]` instead of the contents of
the `index` object.

- To track the interactions on custom components that still update the
values of elements. Moves default element types to `defaultSupportedElements`.
If `extraSupportedElements` set, then it will merge with
`defaultSupportedElements` for a list of supported elements containing
the original elements and user supplied elements.

- `Ga4FinderTracker` previously hardcoded the `type` of the event as
`finder`. If `data-ga4-filter-type` present then it will use this value
instead for `type`.