Allow YT link enhancement when link is added with punctuation
We previously had a check which ensured YouTube link enhancement would only occur when a YT link was on it's own paragraph.

However there's an edge case where a YouTube link is on it's own paragraph, but then has puncutation in the same paragraph outside of the YouTube link.

Link enhancement wasn't running on these links, as the paragraph contained "content" outside of the YT link, so it wouldn't enhance due to our safety checks.

However, punctuation is an OK thing to be replaced with a YT embed, so now we remove punctuation from the checks.