`Ga4EventTracker` use data attributes for schema
If data attribute present on the tracked element that is a schema
attribute, then use this data attribute instead. This is so the
event recorded by `Ga4EventTracker` on click can be updated without
needing to parse, change and then convert back the JSON string used
for setting the event values.