`ReorderableList` tracking on reorder
Track when an item in the list has been interacted with to change the
order of the list. Includes when the user has clicked either of the
buttons or has dragged the item.

If `disable_ga4` not set then we add `ga4-event-tracker` to
`data-module` and `data-ga4-event` attribute to each item.

When the order of the list changes, update `data-index` attribute
so we can accurately send the placement of the interacted item
before it was reordered.

If a user drags and drops a tracked item, then simulate an event
that can be tracked by `ga4-event-tracker` by adding a temporary
data attribute to be sent with a synthetic `click` event.