Add `recordJson` to `Ga4FormTracker`
In more complex forms, just the values of the inputs is not enough to
gain insights into how a user is using the form. Therefore if
`recordJson` is true then the text will be set to a JSON with the keys
of each submitted value being the section that the input is within.