Use `select` component within `select_with_search`
As the select with search component is progressively enhanced, it
renders the select component as a fallback. Before previous changes to
the select component, the select with search was unable to use the
select component directly. Now the select component can be used, the
partial of select with search has been changed to reflect this change.

The select component has been altered to take as a parameter a select
helper.

The `SelectWithSearchHelper` has been altered to delegate other calls
to the `SelectHelper`. It now uses a similar structure to `SelectHelper`
with a `get_options` function. This means it can be used in the select
component partial.