Fix super navigation toggle border styling
Fix an issue where the super navigation menu height shrinks by `1px` when in the open state.

This was caused by the bottom border added to the parent element of the navigation buttons and dropdown menu, when in the open state, the blue border appears at the bottom of the dropdown menu, this gives the appearance of the menu height changing by `1px` as the state is toggled.

The bottom border is now applied to the div element used to constrain the content of the menu, this ensures that the super navigation menu has a constant height of `61px`