Update `aria-labelledby` of `Select with search`
Previously the `aria-labelledby` of the `Select with search` JS
replacement was only being set to the id of the label of the select.
This has been changed to also include the ids of the hint and error if
they are present. We also need to add an id to the label of the `Select`
component for `Select with search` to reference. This is derived from
the `id` parameter (which is required).