Introduce tag component
Replace the gem-c-environment-tag used in layout header with a Tag component based on the Design System component. Moving this to a component gives us advantages such as individual CSS component loading, and validation on what's passed to the component.

A breaking change as content-data-admin has an override class for gem-c-environment-tag which will no longer exist.

Also moves the tag in the header to the top due to the previous design having wrapping issues.

Also adds a border to the GOVUK Frontend print styles for the tag.