Fix broken background colour on `select-with-search` component
The `govuk-colour` mixins were not being parsed in the declarations
at the top of the file. This was because the mixins weren't available
due to `_individual_component_support` being imported later in the
file. Moving that import to the top of the file solves the problem.

Note that this was only noticeable in Specialist Publisher where
the `select-with-search` component appears on a grey background.