Make feedback H3's to H2's to convey correct information hierarchy
Following an issue reported by DAC where the H3's in the feedback component may make those headings appear subordinate to other headings on the page, our accessibility specialist has recommended that the H3's are changed to H2's.

The component does have a H2 but this is hidden when one of the H3 headings on the feedback component is unhidden, which as DAC advised might make the H3 appear related to a previous H2 on the page.

It should be safe to assume that this change will not cause any issues with the existing heading hierarchy on pages where the feedback component appears since the default heading on the component has always been a H2.