Use a template element to build new fieldsets for the add another component
We want to make sure that if the content added to the DOM by the add another component has any JS modules associated with it, that these are initialised when the content is added. To achieve this, we can call `GOVUK.modules.start` after the fieldset has been added.

However, for our select with search module, this didn't work when we were removing the empty fieldset from the DOM and then cloning it and reinstating it to the DOM when the "Add another" button was clicked. It failed because ChoicesJS believed it was already initialised for that fieldset.

To work around this, we use a template element containing the empty fieldset. The template is not part of the DOM, so Choices is not initialised on page load as it is for the empty fieldset in the DOM. Now when the user clicks the add another button, we construct the new fieldset DOM from the template instead of the removed empty fieldset.

The existing empty fieldset rendered to the DOM is retained for cases where JS is disabled.