Remove conditional comments and styles for Internet Explorer
We no longer test in Internet Explorer 8 so the comment targeting it in layout-for-public can be removed.

Only contents-list had .lte-ie8 styles. From a brief bit of testing, it appears that these styles might have not been doing anything anyway anymore on the contents-list focus state, probably because the dash the comment mentions doesn't get rendered by IE as it uses ::before