Add styling for a contact content block
This adds some rudimentary styling to to the Govspeak component for the
Contact Content Block generated via the Content Block Manager. We
originally intended for these styles to be delivered via the
[content_block_tools gem](https://github.com/alphagov/content_block_tools).
However, this resulted in a more weighty CSS output and also meant we
had to pull in a dependency on `govuk-frontend` in another place, which
meant there was a risk of dependencies getting out of whack (see
https://github.com/alphagov/frontend/pull/4943).

After some discussion with @maxgds and @richardTowers, we decided that
the least worse option would be to distribute these styles in the
Components gem. This is a minor misuse of tools, as content blocks are
not _strictly_ a component, but this is already repeating a pattern
that we’ve already established with the packaging of `govspeak` as a
“component but not a component”.