Remove expanded attribute from feedback 'Cancel' buttons
DAC have reported an issue where the two 'Cancel' buttons on the feedback component aren't using the aria-expanded attribute as expected. The buttons do collapse the region, but behave more like a 'Close' button than an expandable control. Users encountering this control as expanded may expect to encounter an associated region after it.

Remove the expanded attribute as per the recommendation and subsequent testing by our accessiblity specialist to resolve the issue.

We also sought feedback from our content designer on whether the button text needs changing to help screen reader users to understand the button's purpose but a decision was made to keep the current button text which was already revised some time ago; the user has clicked on a button to start an action so 'cancel' fits with that action to  'cancel' the journey.

The option to add aria-controls to the buttons instead, to associate the collapsible region with the button was also discussed with our accessibility specialist. But as aria-controls lacks consistent support across assistive tech, a decision was made not to investigate this option any further.

For background on whether the aria-expanded attribute was originally added to the buttons to resolve any specific issue, it appears to have been there a fair while: #163. And this might be the service manual version mentioned on the above PR: https://github.com/alphagov/service-manual-frontend/commits/fff9fed00ba0018e99cac7581bc0853e04ca5e19/app/views/shared/_improve_this_page.html.erb - whilst it doesn't seem to include the aria-expanded attribute, it doesn't seem useful to pursue this any further through the mists of time as we're interested in improving the experience for current screen reader users.