Add `selectCount` option to `Ga4FormTracker`
New option to enable reporting the count of multiple selected options
instead of the text of each selected `option`. This is to overcome an
issue in which very long values were resulting in large `text` fields.

Reworked the documentation for `Ga4FormTracker` as there are now
several options and it needed to be restructured for clarity.