Hide nav menu buttons and links when printing
The GOV.UK logo is now included as a child of the nav element, but this still needs to be included when printing.

This change groups together the navigation elements we want to remove when printing in the CSS, instead of adding the `govuk-!-display-none-print` utility class in several locations throughout the view template.