Fix: ensure links are accessible in IE10
IE10 does not support flexbox, when `display: block` is used the navigation links are moved outside of the navigation header.

Switching to `display: inline-block` ensures the nav links are still accessible in browsers that do not support flexbox.

Removed `float: left` as it is not needed when testing in IE10