Remove the focusable attribute from super nav search close icon
The focusable attribute cannot be used on a span element and is not valid HTML. Using it results in an error when run through a HTML validator: "Attribute focusable not allowed on element span at this point."

It appears that the attribute was added as a part of a wider set of changes to move the super nav focus state to use focus-visible which stops the focus state appearing when a button is clicked on, instead of focusing it by tabbing. I've checked that removing the attribute from the span doesn't result in any changes in the behaviour of focusing the button by clicking or tabbing both in modern browsers and IE11 which doesn't support `focus-visible`.