Add fallback for missing GA4 canonical URL
The canonical url in the head of the page is generated using the domain name and base path. However it only exists on the page if the machine_readable_metadata component is used on the page. Therefore, for pages that don't have this we can recreate the canonical url using window.location.origin and the base_path from the content_item.