Set `row_classes` to be an array
In github.com/alphagov/govuk_publishing_components/pull/4999 we attempted to
address the deprecation errors in the logs. This is mainly fixed, but
there’s still one place where we’re pushing to a string using `<<`. In
other components, we push to arrays, which is supported by `tag.div`,
so for consistency, let’s do it here. This should remove the deprecation
errors in newer Rubies.