Set inverse header background and text colours
Set the default background colour of inverse header to light blue. It will use the standard black text colour so remove setting the inverse option which makes text white on the headings inside the default variant.

Make the html publication header variant the previous darker blue, and move the white text style to this variant.