Change subtext to be used with html publication header
Subtext is currently only used by HTML publications on GOV.UK which use the darker blue Html publication header variant of the component. We're therefore moving the subtext to this variant as the white subtext can't be used with the default light blue inverse header.

We explain this in the documentation and ask developers to add a black subtext as an option to the default inverse header in the future if needed.