Fix unlabelled listbox
During an accessibility review, we were made aware that the `listbox`
element that is created by choices.js did not have an associated label
(required for WCAG 2.2 AA).

This uses the `callbackOnInit` function to add an `aria-labelledby`
attribute that points to the form component’s label, which seems to fix
the issue.