Disabled tests that checked the `checkboxList` height
The option select component includes a CSS class `gem-c-option-select__container`, which sets the max-height to `200px`. This means that even if the JavaScript sets the height to a larger value, `200px` is always returned from the jQuery .height() method, this matches the behaviour of the component in the browser.

We will likely need to refactor the option-select component further and decide if the functionality to set specific heights in JS is required or if this can be removed, following this we can then update the tests as needed.