Hide redundant controls from screen-reader users
The recent DAC report highlighted an issue with the label of the
button that's used to toggle a table between its chart and table view.
The suggested fix was to include context about the table in the button
label itself.

During the attempted fix, I discovered that the premise was incorrect:
the button doesn't need a better label, because it's not needed for
screen-reader users at all. As the chart is already hidden from those
users with the `aria-hidden` attribute, the button has no effect because
the bar chart is never rendered for them. So the button is redundant.

The fix is therefore to hide the button from screen-reader users and
modify the table so that it's only visually hidden, and not completely
hidden by default. This means screen-reader users will experience the
accessible table as intended, while other users can use the toggle
if they want to switch between the two views.