Do not add crest class to organisation logo unless the crest has a visual identity
The crest class causes some extra padding and a border to appear around the organisation logo (or name, if there is no logo). This is not desirable in cases where a custom logo is being used or the organisation has no visual identity.

Previously this has not been a problem, because Whitehall has excluded the crest attribute from the Publishing payload for organisations if the crest has no visual identity or is a custom logo.

However, because we are migrating organisations to Search API v1's govuk index, and collections frontend depends on the "custom" and "no-identity" values in Search API responses to group organisations on taxon pages, we need to start including these crest values in the Publishing API payload.

This means that the organisation logo component must not apply the crest classes when a custom logo is in use or the crest has no identity.