Update related navigation toggle texts for improved accessibility
DAC have flagged that the link text of the Show more toggle buttons in the related nav component was not descriptive enough when the buttons were being navigated to out of context.

To solve this, add visually hidden text, provided by our content designer, to the buttons to make them more descriptive.

Retain the keys `toggle_less` and `toggle_more` that are still being used by the metadata component. The toggle buttons in that component are due to have a similar fix applied, at which point we should be able to remove the said keys, and just use `toggle_less_html` and `toggle_less_html` for both the related nav and metadata toggle buttons.

We've also discussed with the frontenders community how we currently don't have a documented agreed approach for including HTML together with locale keys. We have a ticket in our backlog to decide and document a solution to this, at which point the structure of `toggle_less_html` and `toggle_less_html` might get simplified.