Guess world_location_news paths by document_type
Context: Targets of world_location links can be of the `world_location`
schema, or the `world_location_news` schema. If the target is
a `world_location`, then that document might itself have
a `world_location_news` link to a `world_location_news` document.

Changes: The preferred path to render on the website is constructed from
the base_path of a world_location_news document, otherwise a path is
guessed from the title of a world_location document.

A warning is raised when the path has to be guessed. Arguably it should
raise an error, but that would be a more drastic change, and this
commit's purpose is only to make frontend apps render data from graphql
the same as from content-store.

Previously: Currently govuk_publishing_components doesn't check the
document_type. Instead, it checks whether a base_path is present. This
works for data from the content-api, which checks the document_type of
the link's target during link expansion, and only sends the base_path
when the target is a world_location_news document. But graphql doesn't
check the document_type of the link's target. Graphql used to always
send the base path, until we learned that can lead to differences in the
HTML rendered by frontend apps, so 804ea33 flipped it to never send the
base_path, and now we've noticed that the HTML of other pages is
different. It would be surprisingly hard to get graphql to consider the
document_type of a link's target (we'd have to introduce new types), and
we'd prefer to fix this in the data (a world_location_news path should
always be available) than in the code.

Co-authored-by: Duncan Garmonsway <duncan.garmonsway@digital.cabinet-office.gov.u>