Allow fieldsets to be optionally sortable
This adds a `sortable` option to the `addAnother` component, which
adds move up / move down buttons to fieldsets, allowing the order of
fieldsets to be changed.

This leans on the functionality provided in the (already used elsewhere)
sortable.js library. When a user clicks move up/down, we sort the array
of items and use the `sort` function to sort the fieldsets
programatically.