Ensure focus is maintained when triggered by keyboard
This checks if the move/up buttons are triggered by a keyboard action.
If the target fieldset is at the top, we add focus to the move down
button, if the target fieldset is at the bottom, we add focus to the
move up button, otherwise we add focus to the triggered button.