Allow sortability to work without JS enabled
Although “Add another” does not work without JS enabled, it does show
a minimum of N + 1 items on the page at any one time (where N is the
number of existing items), so a user can still use this component by
submitting a form numerous times. With this in mind, we should ensure
reordering works too.