Update test to use a different slug
Currently the test uses the same slug in the `world_location_news` as
the `world_location`'s name.

This means we aren't properly testing which part of the conditional in
the helper is actually being used.

Updating the test so the slug won't work if we fall back to the `else`
part of the conditional.